PM-38588 - Split ICurrentContextBackfillService into its own file
Match the IUserDecryptionOptionsBuilder / UserDecryptionOptionsBuilder
pattern at the IdentityServer root. The interface keeps its XML docs;
the impl file drops to just the class.
